Add calldata for the w3.labs BatchDeposit contract
This commit adds the erc7730 calldata specification for our BatchDeposit
contract, a permissionless contract for depositing to multiple Ethereum
validators in a single transaction.

The code can be found [on
GitHub](https://github.com/w3labsxyz/ethereum-batch-deposit-contracts).
It has been deployed to Holesky and Mainnet and is verified on
Etherscan.